+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## HEAD
+
+Features:
+- performance: remove jQuery from default autocomplete and checkbox renderers
+- performance: in cell renderers, replace `innerHTML` with `createTextNode`
+
 ## [0.8.16](https://github.com/warpech/jquery-handsontable/tree/v0.8.16) (Mar 26, 2013)
 
 Features: